Implement WorldLineStateEq trait
Define another trait that allows the comparison of the internal quantum
state of the WorldLines. Since stable Rust doesn't allow specialization,
the trait is implemented for every WorldLine object even without internal
states, and will allow the reuse of the swap and open_close updates with
minimal modification to the current logic (and no performance cost).